requirements.txt
blinker==1.4
certifi==2023.11.17
charset-normalizer==3.3.2
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.4.2
dropbox==11.36.2
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
ply==3.11
PyGObject==3.42.1
PyJWT==2.3.0
pymongo==4.6.1
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dotenv==1.0.0
PyYAML==5.4.1
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
stone==3.3.1
systemd-python==234
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.1.0
wadllib==1.3.6
zipp==1.0.0
